完整任务上传
删除了Grade1,Grade2.java文件;
把两个文件合并到新的类MathExam4354;
加入了命令行选择一二年级的代码;
完善了参数合法性的判断：
添加了IO输入，输出到TXT文件的方法。
